+++ rest_framework_services/types/__init__.py
@@ -7,10 +7,12 @@ inputs/outputs are not coupled to the mutation helpers themselves.
 
 from rest_framework_services.types.change_result import ChangeResult
 from rest_framework_services.types.field_change import FieldChange
+from rest_framework_services.types.service_spec import ServiceSpec
 from rest_framework_services.types.unset import UNSET
 
 __all__ = [
     "UNSET",
     "ChangeResult",
     "FieldChange",
+    "ServiceSpec",
 ]